[ADD] Dashboard create edit and delete posts + buttons

+++ routes/posts.js
@@ -7,6 +7,7 @@ module.exports = function(app, session, User, Post, Comment) {
     else
       res.render('new', {
         authenticated: req.session.isAuthenticated, 
+        admin: req.session.admin,
         title: "Create post",
         action: 'create'
       })
@@ -14,12 +15,12 @@ module.exports = function(app, session, User, Post, Comment) {
   
   // Route to render post details
   app.get('/posts/:id', async (req, res) => {
-    const postId = req.params.id;
     const post = await Post.findByPk(req.params.id);
     
     // Pass the 'authenticated' variable to the EJS template
     res.render('show', { 
       authenticated: req.session.isAuthenticated, 
+      admin: req.session.admin,
       post: post,
       title: post.title,
       action: 'show'
@@ -35,6 +36,7 @@ module.exports = function(app, session, User, Post, Comment) {
       if (post) {
         res.render('edit', {
           authenticated: req.session.isAuthenticated,
+          admin: req.session.admin,
           post,
           title: "Edit",
           action: 'edit'
@@ -45,6 +47,7 @@ module.exports = function(app, session, User, Post, Comment) {
   
   // Create Post Route
   app.post('/posts', async (req, res) => {
+    console.log('AHHHH', req.body);
     await Post.create(req.body);
     const lastPost = await Post.findOne({ order: [['id', 'DESC']] });
     res.redirect(`/posts/${lastPost.id}`);
@@ -52,14 +55,21 @@ module.exports = function(app, session, User, Post, Comment) {
   
   // Update Post Route
   app.put('/posts/:id', async (req, res) => {
-    await Post.update(req.body, { where: { id: req.params.id } });
-    res.redirect(`/posts/${req.params.id}`);
+    if (!req.session.admin)
+      res.redirect('/');
+    else {
+      await Post.update(req.body, { where: { id: req.params.id } });
+      res.redirect(`/posts/${req.params.id}`);
+    }
   });
   
   // Delete Post Route
   app.post('/posts/:id', async (req, res) => {
-    console.log('Deleting post with id:', req.params.id);
-    Post.destroy({ where: { id: req.params.id } });
-    res.redirect('/');
+    if (!req.session.admin)
+      res.redirect('/');
+    else {
+      Post.destroy({ where: { id: req.params.id } });
+      res.redirect('/dashboard');
+    }
   });
 }
